Module aws_sdk_detective::error

Modules

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

validation_exception

See ValidationException

Structs

AcceptInvitationError

ConflictException

The request attempted an invalid action.

CreateGraphError

CreateMembersError

DeleteGraphError

DeleteMembersError

DisassociateMembershipError

GetMembersError

InternalServerException

The request was valid but failed because of a problem with the service.

ListGraphsError

ListInvitationsError

ListMembersError

ListTagsForResourceError

RejectInvitationError

ResourceNotFoundException

The request refers to a nonexistent resource.

ServiceQuotaExceededException

This request cannot be completed for one of the following reasons.

StartMonitoringMemberError

TagResourceError

UntagResourceError

ValidationException

The request parameters are invalid.

Enums

AcceptInvitationErrorKind

CreateGraphErrorKind

CreateMembersErrorKind

DeleteGraphErrorKind

DeleteMembersErrorKind

DisassociateMembershipErrorKind

GetMembersErrorKind

ListGraphsErrorKind

ListInvitationsErrorKind

ListMembersErrorKind

ListTagsForResourceErrorKind

RejectInvitationErrorKind

StartMonitoringMemberErrorKind

TagResourceErrorKind

UntagResourceErrorKind