Module aws_sdk_detective::model

Modules

account

See Account

graph

See Graph

member_detail

See MemberDetail

unprocessed_account

See UnprocessedAccount

Structs

Account

An AWS account that is the administrator account of or a member of a behavior

Graph

A behavior graph in Detective.

MemberDetail

Details about a member account that was invited to contribute to a behavior

UnprocessedAccount

A member account that was included in a request but for which the request could not be

Enums

MemberDisabledReason

MemberStatus