Module aws_sdk_detective::operation

Structs

AcceptInvitation

Accepts an invitation for the member account to contribute data to a behavior graph.

CreateGraph

Creates a new behavior graph for the calling account, and sets that account as the

CreateMembers

Sends a request to invite the specified AWS accounts to be member accounts in the

DeleteGraph

Disables the specified behavior graph and queues it to be deleted. This operation

DeleteMembers

Deletes one or more member accounts from the administrator account's behavior graph.

DisassociateMembership

Removes the member account from the specified behavior graph. This operation can only be

GetMembers

Returns the membership details for specified member accounts for a behavior

ListGraphs

Returns the list of behavior graphs that the calling account is an administrator account

ListInvitations

Retrieves the list of open and accepted behavior graph invitations for the member

ListMembers

Retrieves the list of member accounts for a behavior graph. Does not return member

ListTagsForResource

Returns the tag values that are assigned to a behavior graph.

RejectInvitation

Rejects an invitation to contribute the account data to a behavior graph. This operation

StartMonitoringMember

Sends a request to enable data ingest for a member account that has a status of

TagResource

Applies tag values to a behavior graph.

UntagResource

Removes tags from a behavior graph.